LLM_Assignment/LLM_GPT_04/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
Django==4.2
django-extensions==3.2.3
django-redis==5.4.0
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Faker==33.3.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
mysqlclient==2.2.7
openai==0.28.0
pillow==11.1.0
propcache==0.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
toposort==1.10
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
